fix: 修复 P2 阶段 3 个后端 bug + 恢复 Android .gitignore
BUG-001: WorkoutReminderWorker 通知 ID 碰撞 — 改用 planId 生成通知 ID，
  避免 System.currentTimeMillis().toInt() 同毫秒碰撞
BUG-002: PressureAnalyzer 疲劳压力整数除法精度丢失 — 改为浮点除法再取整
BUG-003: QwenRepository extractJson 脆弱提取 — 增加括号匹配 + markdown
  代码块解析，正确处理嵌套 JSON 和转义字符
fix: .gitignore 被覆盖为 Node.js 模式，恢复 Android 标准配置

+++ app/src/main/java/com/fittrack/data/analyzer/PressureAnalyzer.kt
@@ -192,7 +192,7 @@ class PressureAnalyzer {
             val lowEnergyCount = recentRecords.count { it.energyLevel <= 2 }
             val lowSleepCount = recentRecords.count { it.sleepQuality <= 2 }
 
-            return ((lowEnergyCount * 20 + lowSleepCount * 25) / recentRecords.size).toInt()
+            return ((lowEnergyCount * 20 + lowSleepCount * 25).toFloat() / recentRecords.size).toInt()
                 .coerceIn(0, 100)
         }
